Big changes, all branches were combine into one. Change log is too big, see the file changes
