+++ go.mod
@@ -3,6 +3,7 @@ module github.com/rabbytesoftware/quiver
 go 1.24.2
 
 require (
+	github.com/gorilla/mux v1.8.1
 	github.com/huin/goupnp v1.3.0
 	github.com/pterm/pterm v0.12.80
 	github.com/rabbytesoftware/quiver.compiler v1.25.5
@@ -14,7 +15,6 @@ require (
 	atomicgo.dev/schedule v0.1.0 // indirect
 	github.com/containerd/console v1.0.3 // indirect
 	github.com/gookit/color v1.5.4 // indirect
-	github.com/gorilla/mux v1.8.1 // indirect
 	github.com/lithammer/fuzzysearch v1.1.8 // indirect
 	github.com/mattn/go-runewidth v0.0.16 // indirect
 	github.com/rivo/uniseg v0.4.7 // indirect
