+++ logs/fatal.txt
@@ -1,0 +1,3 @@
+[fatal] init: Failed to discover packages: failed to read packages directory: open ./pkgs: no such file or directory - (2025-05-23 22:22:09)
+[fatal] init: Failed to discover packages: failed to read packages directory: open ./pkgs: no such file or directory - (2025-05-23 22:22:59)
+[fatal] init: Failed to discover packages: failed to read packages directory: open ./pkgs: no such file or directory - (2025-05-23 22:23:23)
