+++ logs/info.txt
@@ -1,0 +1,42 @@
+[info] welcome:  _____       _                 - (2025-05-23 22:21:59)
+[info] welcome: |  _  |     (_)                - (2025-05-23 22:21:59)
+[info] welcome: | | | |_   _ ___   _____ _ __  - (2025-05-23 22:21:59)
+[info] welcome: | | | | | | | \ \ / / _ \ '__| - (2025-05-23 22:21:59)
+[info] welcome: \ \/' / |_| | |\ V /  __/ |    - (2025-05-23 22:21:59)
+[info] welcome:  \_/\_\\__,_|_| \_/ \___|_|    - (2025-05-23 22:21:59)
+[info] welcome:                                - (2025-05-23 22:21:59)
+[info] welcome: ~~~~~~~~~ Quiver SDK ~~~~~~~~~ - (2025-05-23 22:21:59)
+[info] welcome: Developed by the people behind Rabbyte Software. - (2025-05-23 22:21:59)
+[info] welcome: Quiver and Quiver SDK is license under GPL-3.0 license. - (2025-05-23 22:21:59)
+[info] welcome: Version: dev 6.25 - (2025-05-23 22:21:59)
+[info] welcome: Maintainers: @char2cs, @AgustinGil21 & @Lucacux - (2025-05-23 22:21:59)
+[info] welcome:  - (2025-05-23 22:21:59)
+[info] welcome:  - (2025-05-23 22:21:59)
+[info] welcome:  _____       _                 - (2025-05-23 22:22:49)
+[info] welcome: |  _  |     (_)                - (2025-05-23 22:22:49)
+[info] welcome: | | | |_   _ ___   _____ _ __  - (2025-05-23 22:22:49)
+[info] welcome: | | | | | | | \ \ / / _ \ '__| - (2025-05-23 22:22:49)
+[info] welcome: \ \/' / |_| | |\ V /  __/ |    - (2025-05-23 22:22:49)
+[info] welcome:  \_/\_\\__,_|_| \_/ \___|_|    - (2025-05-23 22:22:49)
+[info] welcome:                                - (2025-05-23 22:22:49)
+[info] welcome: ~~~~~~~~~ Quiver SDK ~~~~~~~~~ - (2025-05-23 22:22:49)
+[info] welcome: Developed by the people behind Rabbyte Software. - (2025-05-23 22:22:49)
+[info] welcome: Quiver and Quiver SDK is license under GPL-3.0 license. - (2025-05-23 22:22:49)
+[info] welcome: Version: dev 6.25 - (2025-05-23 22:22:49)
+[info] welcome: Maintainers: @char2cs, @AgustinGil21 & @Lucacux - (2025-05-23 22:22:49)
+[info] welcome:  - (2025-05-23 22:22:49)
+[info] welcome:  - (2025-05-23 22:22:49)
+[info] welcome:  _____       _                 - (2025-05-23 22:23:13)
+[info] welcome: |  _  |     (_)                - (2025-05-23 22:23:13)
+[info] welcome: | | | |_   _ ___   _____ _ __  - (2025-05-23 22:23:13)
+[info] welcome: | | | | | | | \ \ / / _ \ '__| - (2025-05-23 22:23:13)
+[info] welcome: \ \/' / |_| | |\ V /  __/ |    - (2025-05-23 22:23:13)
+[info] welcome:  \_/\_\\__,_|_| \_/ \___|_|    - (2025-05-23 22:23:13)
+[info] welcome:                                - (2025-05-23 22:23:13)
+[info] welcome: ~~~~~~~~~ Quiver SDK ~~~~~~~~~ - (2025-05-23 22:23:13)
+[info] welcome: Developed by the people behind Rabbyte Software. - (2025-05-23 22:23:13)
+[info] welcome: Quiver and Quiver SDK is license under GPL-3.0 license. - (2025-05-23 22:23:13)
+[info] welcome: Version: dev 6.25 - (2025-05-23 22:23:13)
+[info] welcome: Maintainers: @char2cs, @AgustinGil21 & @Lucacux - (2025-05-23 22:23:13)
+[info] welcome:  - (2025-05-23 22:23:13)
+[info] welcome:  - (2025-05-23 22:23:13)
