+++ logs/ok.txt
@@ -1,0 +1,3 @@
+[ok] init: Device registered with IP: 179.41.9.186 - (2025-05-23 22:22:09)
+[ok] init: Device registered with IP: 179.41.9.186 - (2025-05-23 22:22:59)
+[ok] init: Device registered with IP: 179.41.9.186 - (2025-05-23 22:23:23)
