--- rest/shared/utils/packages/global-variables/global-variables.go
@@ -1,7 +1,0 @@
-package packages_global_variables
-
-import (
-	pkgs "rounds.com.ar/watcher/packages"
-)
-
-var Packages map[string]*pkgs.Package
