--- view/logger/log-files-minor-utilities.go
@@ -1,245 +1,0 @@
-package logger
-
-import (
-	"compress/gzip"
-	"fmt"
-	"io"
-	"os"
-	"path/filepath"
-	"strings"
-	"regexp"
-)
-
-func CopyFileContent(originalFile, newFile *os.File) error {
-	// Move the read pointer to the
-	// beginning of the original file
-	_, seekErr := originalFile.Seek(0, 0)
-	if seekErr != nil {
-		return fmt.Errorf("could not seek file: %w", seekErr)
-	}
-
-	// Copy content
-	_, copyErr := io.Copy(newFile, originalFile)
-
-	if copyErr != nil {
-		return fmt.Errorf("could not copy content: %w", copyErr)
-	}
-
-	return nil
-}
-
-func CopyFileContentToCompressedGzipFile(originalFile *os.File, compressedFile *gzip.Writer) error {
-	// Move the read pointer to the
-	// beginning of the original file
-	_, seekErr := originalFile.Seek(0, 0)
-	if seekErr != nil {
-		return fmt.Errorf("could not seek file: %w", seekErr)
-	}
-
-	// Open original file as
-	// reader
-	originalFilePath := GetFilePath(originalFile)
-
-	readOriginalPath, readErr := os.Open(originalFilePath)
-
-	if readErr != nil {
-		return fmt.Errorf("could not read file %s", originalFilePath)
-	}
-
-	// Copy content
-	_, copyErr := io.Copy(compressedFile, readOriginalPath)
-
-
-	if copyErr != nil {
-		return fmt.Errorf("could not copy content: %w", copyErr)
-	}
-
-	// Close compressed file
-	if err := compressedFile.Close(); err != nil {
-		return fmt.Errorf("could not close compressed file: %w", err)
-	}
-
-	return nil
-}
-
-func DeleteFile(f *os.File) error {
-	filePath := GetFilePath(f)
-
-	// Attempt to close the file before deleting it
-	if closeErr := f.Close(); closeErr != nil {
-		return fmt.Errorf("failed to close file: %w", closeErr)
-	}
-
-	// Attempt to delete the file after closing
-	if deleteErr := os.Remove(filePath); deleteErr != nil {
-		return fmt.Errorf("failed to delete file: %w", deleteErr)
-	}
-
-	return nil
-}
-
-func CompressToGzipFile(f *os.File) *gzip.Writer {
-	gzWriter := gzip.NewWriter(f)
-
-	return gzWriter
-}
-
-func FindFile(root, filename string) (string, error) {
-	var foundPath string
-
-	// Search by filename
-	err := filepath.WalkDir(root, func(path string, d os.DirEntry, err error) error {
-    if err != nil {
-      return err
-    }
-		
-    if !d.IsDir() && d.Name() == filename {
-      foundPath = path
-      return filepath.SkipDir // stop search
-    }
-    
-		return nil
-  })
-
-	// Check if error
-  if err != nil {
-    return "", err
-  }
-
-	// Check if path is empty
-  if foundPath == "" {
-    return "", fmt.Errorf("file %s not found", filename)
-  }
-  
-	return foundPath, nil
-}
-
-func CheckIfFileExists(root, filename string) bool {
-	_, err := FindFile(root, filename)
-
-	return err != nil
-}
-
-func GetFile(root, filename string) (*os.File, error) {
-	// Get file path
-	filePath, findErr := FindFile(root, filename)
-
-	if findErr != nil {
-		return nil, fmt.Errorf("file not found: %w", findErr)
-	}
-
-	// Get file
-	file, openErr := os.OpenFile(filePath, os.O_WRONLY|os.O_APPEND|os.O_CREATE, 0644)
-
-	if openErr != nil {
-		return nil, fmt.Errorf("could not open file: %w", openErr)
-	}
-
-	return file, nil
-}
-
-func GetFilePath(f *os.File) string {
-	return f.Name()
-}
-
-func GetFilename(f *os.File) string {
-	path := GetFilePath(f)
-
-	pathParts := strings.Split(path, "/")
-
-	// Get last one (filename)
-	return pathParts[len(pathParts) - 1]
-}
-
-func CreateFolder(root, name string) error {
-	// Create folder if not exists
-	folderPath := fmt.Sprintf("%s/%s", root, name)
-	err := os.MkdirAll(folderPath, os.ModePerm)
-
-	if err != nil {
-		return fmt.Errorf("could not create folder: %w", err)
-	}
-	
-	return nil
-}
-
-func RepeatedSameFiles(path, name string) (bool, error) {
-	// List files in path
-	files, filesErr := ListFiles(path)
-
-	if filesErr != nil {
-		return false, filesErr
-	}
-
-	// Check if file exists
-	for _, file := range files {
-		// Separators
-		re := regexp.MustCompile(`[-_.]`)
-		splitFile := re.Split(file, -1)
-
-		if strings.EqualFold(splitFile[0], name) {
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
-func ListFiles(path string) ([]string, error) {
-	// List dir entries
-	entries, err := os.ReadDir(path)
-	if err != nil {
-		return nil, fmt.Errorf("could not read directory: %w", err)
-	}
-
-	var files []string
-
-	// Check if entry is a file
-	for _, entry := range entries {
-		if !entry.IsDir() {
-			files = append(files, entry.Name())
-		}
-	}
-
-	return files, nil
-}
-
-func ListFolders(path string) ([]string, error) {
-	// List dir entries
-	entries, err := os.ReadDir(path)
-	if err != nil {
-		return nil, fmt.Errorf("could not read directory: %w", err)
-	}
-
-	var folders []string
-
-	// Check if entry is a folder
-	for _, entry := range entries {
-		if entry.IsDir() {
-			folders = append(folders, entry.Name())
-		}
-	}
-
-	return folders, nil
-}
-
-func CheckIfFolderExists(root, name string) (bool, error) {
-	// Get folders list
-	folders, foldersErr := ListFolders(root)
-
-	if foldersErr != nil {
-		return false, foldersErr
-	}
-
-	// Check if folder exists
-	for _, folderName := range folders {
-		if strings.EqualFold(folderName, name) {
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
-
-
